refactor(settings): settings_dialog als Paket — dialog.py + oauth_task.py + Re-Exports (H4 Schritt 1)
